compiler: frame level mismatch bug.
Repro test case:

  (defvar v)
  (defun f (: (v v)))
  (call (compile 'f)) ;; blows up

* share/txr/stdlib/compiler.tl (compiler comp-lambda): The
specials variable is wrongly used as a Boolean to decide
whether we need an extra environment level for specials.
The problem occurs when all of the specials are optional
parameters. Optionals are handled specially and removed from
the specials list. If all the specials are optional
parameters, then the special list becomes empty. This cannot
be intepreted as "there are no specials". The correct Boolean
which indicates "there are specials" is need-dframe.