defun: record source loc info.
The interpreted defun should tag the function form that is
obtainable with func-get-form with source location info.

* eval.c (op_defun): In all cases, propagate source loc info
from the form to the function form that is bound to
the function name.
(op_defmacro): Propagate the source location info in the
same manner: not from the body (which could be empty) but from
the form.