replace_list: revise, fixing several bugs.
* lib.c (replace_list): Avoid using the dubious toseq
function which turns a non-sequence object into a list of one
item. This is not documented. Rewrite the
assign-to-multiple-indices logic into one loop that uses
sequence iterators. A bug is fixed here: failing to step the
indices and items in parallel.  Don't use list_vec on items;
it fails on strings. Strings are now supported properly: they
expand into a list of items that is spliced.