structs: gc bug due to derived hook.
* struct.c (make_struct_type): The static slot lookup for for
a derived method in the supertype can potentially allocate a
cons cell when there is a cache miss, which can trigger gc. So
we can't be doing this while a the new type object object is
partially initialized, because then the garbage collector will
visit uninitialized memory. We move the derived method
lookup down farther to where that value is needed to do the
derived call, just before we return the new type.