packages: generational gc bug.
* lib.c (make_package_common): The way the two hashes are
assigned into the new package here is not correct. The problem
is that the first make_hash can trigger gc. Then it is
possible that the package object will move into the mature
generation, after which the assignment of the second package
is a wrong-way assignment requiring the set macro.
Instead of bringing in that macro, the obvious way to solve it
is to just allocate the hashes first, and then the package:
exactly the way we build a cons cell from existing values.