seq_info: nullify bugfix.
A change in the nullify function to support hash tables has
broken various functions which classify an object using
seq_info, obtainig a SEQ_HASHLIKE kind, and then work with
si.obj using hash functions. But si.obj has been nullified.
An example of a broken function is find-max.

Basically, this can be attributed to a careless use of nullify
in seq_info. The purpose of nullify is to support code which
treats any sequence as if it were a list. But seq_info doesn't
do that; it classifies sequences and treats them according to
their kind. Under seq_info, the only non-list objects that get
treated as lists are list-like structures. For these it makes
sense to call nullify, in case they have a nullify method.

* lib.c (seq_info): Don't unconditionally call nullify on all
COBJ objects. Only call nullify on struct objects. If that
returns nil, then treat the object as SEQ_NIL; and if it
returns an object different from the original, then recurse.