expander: bugfix: spurious lambda form warning.
* eval.c (do_expand): In the expansion logic for function
calls, after we have done the dot-to-apply transform,
we must thereafter consistently refer to the new front element
of the form insym, and not the original element sym.
The first element may be a lambda form moved into the second
position by dot-to-apply. We then falsely warn about that
being in the operator position. Test case: ((lambda ()) . 3).