parser: bugfix: uninitialized ignore flag.
* parser.c (parser_common_init): Initialize the ignore flag,
which was until recently called circ_suppress.  When the
parser is invoked via parse_once or parse_once_noerror, rather
than parse, the state of the flag is indeterminate. Thus this
only affects the TXR Pattern Language, not TXR Lisp. As a
result of this bug, which affects releases 157 through 223,
if the circle notation like #1=(foo #1#) is used in Pattern
Language syntax, it may not be handled properly.  I discovered
this bug now because there are new behaviors connected to the
flag; it doesn't just affect the processing of the circle
notation, but is involved in all symbol handling in the
parser.