doc: fixes from vapnik spaknik.
TXR user vapnik spaknik has found a swath of errors in the
documentation.

* txr.1: Fixed bugs in numerous examples, copy-and-paste
errors in syntax sections, and other typos, including *stdin*
mistyped as *std-input* in many places.  A new executable code
example is provided for expand-with-free-refs to calculate the
result that was previously only verbally described.  Fixed
*stdin* mistyped as *std-input* (a non-existent symbol that
imitates the internal C variable).