list-builder: bugfix: broken self-appending.
The list builder is failing on the documented example

  (build
    (add 1 2)
    (pend (get))
    (pend (get)))
  -> (1 2 1 2 1 2 1 2)

wrongly constructing an infinite list.

* share/txr/stdlib/build.tl (list-builder pend): When
destructively appending the next argument, check whether
the current tail is a tail of that object. If so, copy
the object to prevent a cycle from forming.
(list-builder pend*): When appending the old head to the
catenated list, do the tail check and copy the object if
necessary to prevent the creation of a cycle.