Add forgotten alloca header under version control.
I had this file locally for the better part of a year!
Commit 5e4c74dfd5927b3829b4f5e04a7964dbac6a4f34 made on
December 31, 2018 claims that alloca.h was added.
In fact, it was not. Things work without it because
the #include "alloca.h" directives in the code end up
resolving to <alloca.h>. This is widely available so nobody
has seen a problem.

* alloca.h: New file.