safety: fix type tests that code can subvert.
This patch fixes numerous instances of a safety hole which
involves the type of a COBJ object being tested to be of
a given class using logic that can be subverted by the
definition of a like-named struct.

Specifically logic like (typeof(obj) == hash_s) is broken,
because if a struct type called hash is defined, then the test
will yield true for instances of that struct type. Those
instances can then be passed into code that only works on COBJ
hashes, and relies on this test to reject invalid objects.

* ffi.c (make_carray): Replace fragile test with strong one,
using new cobjclassp function.

* hash.c (hashp): Likewise.

* lib.c (class_check): The expression used here for the type
test moves into the new function cobjclassp and so is replaced
by a call to that function.
(cobjclassp): New function.

* lib.h (cobjclassp): Declared.

* rand.c (random_state_p): Replace fragile test using
cobjclassp.

* regex.c (char_set_compile): Replace fragile typeof tests
for character type with is_chr.
(reg_derivative, regexp): Replace fragile test with cobjclassp.

* struct.c (struct_type_p): Replace fragile test with cobjclassp.